feat: move wrapper-types to separate file
