+++ src/components/BaseTable/__stories__/stories/ColumnPinningStory.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import type {ColumnPinningState} from '@tanstack/react-table';
 
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import {BaseTable} from '../../BaseTable';
 import {ColumnPinningHeaderCell} from '../cells/ColumnPinningHeaderCell';
 import {data} from '../constants/data';
