+++ src/components/BaseTable/__stories__/stories/CustomRowStory.tsx
@@ -1,7 +1,7 @@
 import type {Row} from '@tanstack/react-table';
 
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import type {BaseTableProps} from '../../BaseTable';
 import {BaseTable} from '../../BaseTable';
 import {data} from '../constants/data';
