+++ src/components/BaseTable/__stories__/stories/ReorderingTreeStory.tsx
@@ -4,7 +4,7 @@ import type {ExpandedState} from '@tanstack/react-table';
 
 import {dragHandleColumn} from '../../../../constants';
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import {ReorderingProvider} from '../../../ReorderingProvider';
 import {BaseTable} from '../../BaseTable';
 import type {TreeItem} from '../constants/tree';
