+++ src/components/RowActions/__stories__/stories/RowActionsWithActionsColumnStory.tsx
@@ -4,7 +4,7 @@ import type {ColumnPinningState} from '@tanstack/react-table';
 
 import {SETTINGS_WITH_ACTIONS_COLUMN_ID, getSettingsWithActionsColumn} from '../../../../constants';
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import type {Item} from '../../../BaseTable/__stories__/types';
 import {generateData} from '../../../BaseTable/__stories__/utils';
 import {Table} from '../../../Table/Table';
