+++ src/components/Table/__stories__/constants/grouping.tsx
@@ -1,4 +1,4 @@
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import type {Item} from '../../../BaseTable/__stories__/types';
 import {TreeExpandableCell} from '../../../TreeExpandableCell';
