+++ src/components/Table/__stories__/stories/ReorderingStory.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import {dragHandleColumn} from '../../../../constants';
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import {columns as originalColumns} from '../../../BaseTable/__stories__/constants/columns';
 import {data as originalData} from '../../../BaseTable/__stories__/constants/data';
 import type {Item} from '../../../BaseTable/__stories__/types';
