+++ src/components/Table/__stories__/stories/WithSelectionStory.tsx
@@ -4,7 +4,7 @@ import type {RowSelectionState} from '@tanstack/react-table';
 
 import {selectionColumn} from '../../../../constants';
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import {columns as originalColumns} from '../../../BaseTable/__stories__/constants/columns';
 import {data} from '../../../BaseTable/__stories__/constants/data';
 import type {Item} from '../../../BaseTable/__stories__/types';
