+++ src/components/TableSettings/TableSettings.extentions.ts
@@ -1,6 +1,6 @@
 import type {RowData} from '@tanstack/react-table';
 
-import type {Column} from '../../types/tanstack';
+import type {Column} from '../../types/base';
 
 declare module '@tanstack/react-table' {
     interface ColumnMeta<TData extends RowData, TValue> {
