+++ src/components/TableSettings/TableSettings.utils.ts
@@ -3,7 +3,7 @@ import * as React from 'react';
 import type {DragEndEvent, DragStartEvent, UniqueIdentifier} from '@dnd-kit/core';
 import {arrayMove} from '@dnd-kit/sortable';
 
-import type {Column} from '../../types/tanstack';
+import type {Column} from '../../types/base';
 
 const filterColumns = <TData extends unknown>(
     column: Column<TData> | undefined,
