+++ src/components/TableSettings/__stories__/stories/TableSettingsColumnStory.tsx
@@ -4,7 +4,7 @@ import type {ColumnPinningState, RowSelectionState} from '@tanstack/react-table'
 
 import {SETTINGS_COLUMN_ID, getSettingsColumn, selectionColumn} from '../../../../constants';
 import {useTable} from '../../../../hooks';
-import type {ColumnDef} from '../../../../types/tanstack';
+import type {ColumnDef} from '../../../../types/base';
 import {Table} from '../../../Table/Table';
 import type {TableSettingsOptions} from '../../TableSettings';
