+++ src/constants/actionsColumn.tsx
@@ -1,6 +1,6 @@
 import {ActionsCell} from '../components';
 import type {TableActionsSettings} from '../types/RowActions';
-import type {ColumnDef} from '../types/tanstack';
+import type {ColumnDef} from '../types/base';
 
 export const ACTIONS_COLUMN_ID = '_actions';
 const ACTIONS_COLUMN_SIZE = 44;
