+++ src/constants/settingsWithActionsColumn.ts
@@ -1,6 +1,6 @@
 import type {TableSettingsOptions} from '../components';
 import type {TableActionsSettings} from '../types/RowActions';
-import type {ColumnDef} from '../types/tanstack';
+import type {ColumnDef} from '../types/base';
 
 import {getActionsColumn} from './actionsColumn';
 import {getSettingsColumn} from './settingsColumn';
