+++ src/index.ts
@@ -38,7 +38,6 @@ export {
 export type {
     UseDraggableRowDepthOptions,
     UseRowVirtualizerOptions,
-    UseTableOptions,
     UseWindowRowVirtualizerOptions,
     UseTableSettingsOptions,
 } from './hooks';
@@ -55,3 +54,5 @@ export type {
 } from './types/RowActions';
 
 export type {TableSize} from './components/Table/types';
+
+export type * from './types/base';
