+++ src/tanstack.ts
@@ -1,3 +1,1 @@
 export * from '@tanstack/react-table';
-
-export type {ColumnDef, Column, Cell, Header} from './types/tanstack';
