+++ src/utils/getCellClassModes.ts
@@ -1,6 +1,6 @@
 import type {NoStrictEntityMods} from '@bem-react/classname';
 
-import type {Cell} from '../types/tanstack';
+import type {Cell} from '../types/base';
 
 import {getColumnPinningClassModes} from './getColumnPinningClassModes';
