+++ src/utils/getHeaderCellAriaColIndex.ts
@@ -1,4 +1,4 @@
-import type {Header} from '../types/tanstack';
+import type {Header} from '../types/base';
 
 export const getHeaderCellAriaColIndex = <TData, TValue = unknown>(
     header: Header<TData, TValue>,
