+++ src/utils/shouldRenderFooterCell.ts
@@ -1,4 +1,4 @@
-import type {Header} from '../types/tanstack';
+import type {Header} from '../types/base';
 
 export const shouldRenderFooterCell = <TData>(header: Header<TData, unknown>) => {
     return !header.isPlaceholder;
