+++ src/utils/shouldRenderFooterRow.ts
@@ -1,4 +1,4 @@
-import type {HeaderGroup} from '@tanstack/react-table';
+import type {HeaderGroup} from '../types/base';
 
 export const shouldRenderFooterRow = <TData>(footerGroup: HeaderGroup<TData>) => {
     return footerGroup.headers.some((header) => header.column.columnDef.footer);
